+++ R/02-item_methods.R
@@ -2377,7 +2377,7 @@ setClass('custom', contains = 'AllItemsClass',
                                          hss='function',
                                          gen='function',
                                          usehss='logical',
-                                         userdata='matrix',
+                                         userdata='list',
                                          useuserdata='logical',
                                          derivType='character',
                                          derivType.hss='character'))
@@ -2436,15 +2436,15 @@ setMethod(
     f = "ProbTrace",
     signature = signature(x = 'custom', Theta = 'matrix'),
     definition = function(x, Theta, useDesign = TRUE, ot=0){
-        if(x@useuserdata) Theta <- cbind(Theta, x@userdata)
-        return(x@P(x@par, Theta=Theta, ncat=x@ncat))
+        if(x@useuserdata) return(x@P(x@par, Theta, x@ncat, x@userdata[[1L]]))
+        else return(x@P(x@par, Theta, x@ncat))
     }
 )
 
 setMethod("initialize",
           'custom',
-          function(.Object, name, par, est, parnames, lbound, ubound, P, gr, hss, gen, userdata, derivType,
-                   derivType.hss, dps, dps2) {
+          function(.Object, name, par, est, parnames, lbound, ubound,
+                   P, gr, hss, gen, userdata, derivType, derivType.hss, dps, dps2) {
               dummyfun <- function(...) return(NULL)
               names(est) <- names(par)
               usegr <- usehss <- useuserdata <- TRUE
@@ -2470,14 +2470,14 @@ setMethod("initialize",
                   .Object@gen <- function(object) object@par
               } else .Object@gen <- gen
               if(is.null(userdata)){
-                  .Object@userdata <- matrix(NaN)
+                  .Object@userdata <- list()
                   useuserdata <- FALSE
               } else .Object@userdata <- userdata
               .Object@usegr <- usegr
               .Object@usehss <- usehss
               .Object@useuserdata <- useuserdata
-              .Object@lbound <- if(!is.null(lbound)) lbound  else rep(-Inf, length(par))
-              .Object@ubound <- if(!is.null(ubound)) ubound  else rep(Inf, length(par))
+              .Object@lbound <- if(!is.null(lbound)) lbound else rep(-Inf, length(par))
+              .Object@ubound <- if(!is.null(ubound)) ubound else rep(Inf, length(par))
               .Object
           }
 )
@@ -2486,7 +2486,6 @@ setMethod(
     f = "Deriv",
     signature = signature(x = 'custom', Theta = 'matrix'),
     definition = function(x, Theta, estHess = FALSE, offterm = numeric(1L)){
-        if(x@useuserdata) Theta <- cbind(Theta, x@userdata)
         grad <- rep(0, length(x@par))
         hess <- matrix(0, length(x@par), length(x@par))
         if(x@usegr) grad <- x@gr(x, Theta)