+++ R/PrepData.R
@@ -1,8 +1,8 @@
 PrepData <- function(data, model, itemtype, guess, upper, gpcm_mats, opts,
                      parprior, verbose, technical, parnumber = 1, BFACTOR = FALSE,
                      grsm.block = NULL, rsm.block = NULL, mixed.design, customItems,
-                     customGroup, fulldata = NULL, key, spline_args, internal_constraints,
-                     monopoly.k, dentype, dcIRT_nphi)
+                     customItemsData, customGroup, fulldata = NULL, key,
+                     spline_args, internal_constraints, monopoly.k, dentype, dcIRT_nphi)
 {
     if(is.null(grsm.block)) grsm.block <- rep(1, ncol(data))
     if(is.null(rsm.block)) rsm.block <- rep(1, ncol(data))
@@ -103,7 +103,8 @@ PrepData <- function(data, model, itemtype, guess, upper, gpcm_mats, opts,
                            itemloc=itemloc, data=data, N=N, guess=guess, upper=upper,
                            itemnames=itemnames, exploratory=exploratory, parprior=parprior,
                            parnumber=parnumber, BFACTOR=BFACTOR, mixed.design=mixed.design,
-                           customItems=customItems, customGroup=customGroup, key=key,
+                           customItems=customItems, customItemsData=customItemsData,
+                           customGroup=customGroup, key=key,
                            gpcm_mats=gpcm_mats, spline_args=spline_args, monopoly.k=monopoly.k,
                            dcIRT_nphi=dcIRT_nphi, dentype=dentype)
     prodlist <- attr(pars, 'prodlist')