+++ R/createItem.R
@@ -2,8 +2,12 @@
 #'
 #' Initializes the proper S4 class and methods necessary for mirt functions to use in estimation.
 #' To use the defined objects pass to the \code{mirt(..., customItems = list())} command, and
-#' ensure that the classes are properly labeled and unique in the list. For further examples
-#' regarding how this function can be used for fitting unfolding-type models see Liu and Chalmers (2018).
+#' ensure that the classes are properly labeled and unique in the list. Additionally,
+#' the input \code{mirt(..., customItemsData = list())} can also be included to specify additional
+#' item-level information to better recycle custom-item defintions (e.g., for supplying varying
+#' Q-matricies), where the \code{list} input must have the same length as the number of items.
+#' For further examples regarding how this function can be used for
+#' fitting unfolding-type models see Liu and Chalmers (2018).
 #'
 #' The \code{summary()} function will not return proper standardized loadings since the function
 #' is not sure how to handle them (no slopes could be defined at all!). Instead loadings of .001
@@ -15,8 +19,20 @@
 #' @param est a logical vector indicating which parameters should be freely estimated by default
 #' @param P the probability trace function for all categories (first column is category 1, second
 #'   category two, etc). First input contains a vector of all the item parameters, the second input
-#'   must be a matrix called \code{Theta}, and the third input must be the number of categories
-#'   called \code{ncat}. Function also must return a \code{matrix} object of category probabilities
+#'   must be a matrix called \code{Theta}, the third input must be the number of categories
+#'   called \code{ncat}, and (optionally) a fourth argument termed \code{itemdata}
+#'   may be included containing further  users specification information.
+#'   The last optional input is to be utilized within the estimation functions
+#'   such as \code{\link{mirt}} via the list input \code{customItemsData}}
+#'   to more naturally recycle custom-item definitions. Therefore, these inputs must be of the form
+#'
+#'   \code{function(par, Theta, ncat){...}}
+#'   or
+#'   \code{function(par, Theta, ncat, itemdata){...}}
+#'   to be valid; however, the names of the arguements is not relavent.
+#'
+#'   Finally, this function must return a \code{matrix} object of category probabilities, where
+#'   the columns represent each respective category
 #' @param gr gradient function (vector of first derivatives) of the log-likelihood used in
 #'   estimation. The function must be of the form \code{gr(x, Theta)}, where \code{x} is the object
 #'   defined by \code{createItem()} and \code{Theta} is a matrix of latent trait parameters.
@@ -170,6 +186,9 @@ createItem <- function(name, par, est, P, gr=NULL, hss = NULL, gen = NULL,
     if(missing(par)) missingMsg('par')
     if(missing(est)) missingMsg('est')
     if(missing(P)) missingMsg('P')
+    nms_args <- names(formals(P))
+    if(!(length(nms_args) %in% c(3L,4L)))
+        stop('P function does not have the correct number of arguments', call.=FALSE)
     if(any(names(par) %in% c('g', 'u')) || any(names(est) %in% c('g', 'u')))
         stop('Parameter names cannot be \'g\' or \'u\', please change.', call.=FALSE)
     if(bytecompile) P <- compiler::cmpfun(P)
@@ -189,7 +208,10 @@ createItem <- function(name, par, est, P, gr=NULL, hss = NULL, gen = NULL,
         hss <- symbolicHessian_par
     }
     if(bytecompile && !is.null(hss)) hss <- compiler::cmpfun(hss)
-    return(new('custom', name=name, par=par, est=est, parnames=names(par), lbound=lbound,
+    ret <- new('custom', name=name, par=par, est=est, parnames=names(par), lbound=lbound,
                ubound=ubound, P=P, dps=dps, dps2=dps2, gr=gr, hss=hss, gen=gen, userdata=NULL,
-               derivType=derivType, derivType.hss=derivType.hss))
+               derivType=derivType, derivType.hss=derivType.hss)
+    ret@useuserdata <- if(length(nms_args) == 3L) FALSE else TRUE
+    ret
+
 }